move pair details secondary buttons into info summary
